// src/enums/headers.type.enum.ts
export enum Referer {
  AZTEC = "https://aztecscan.xyz/",
  DASHTEC = "https://www.dashtec.xyz/",
  NETHERMIND = "https://aztec.nethermind.io/",
}

export enum HeadersProperties {
  ACCEPT = "*/*",
  ACCEPT_LANGUAGE = "en-US,en;q=0.9",
  CONNECTION = "keep-alive",
  SEC_FETCH_DEST = "empty",
  SEC_FETCH_MODE = "cors",
  SEC_FETCH_SITE = "same-origin",
  TE = "trailers",
}